requirements.txt
streamlit==1.24.0
sentence-transformers==2.2.2
transformers==4.33.2
Pillow==9.4.0
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.2
torch==2.0.1
